requirements.txt
tk
Pillow